refactor: Rename fastapi_dependencies -> fastapi.
Since we're also going to have exception handlers in here now, not just dependencies.
