+++ auth-server/auth_server/users/router.py
@@ -2,7 +2,7 @@ from typing import Annotated
 
 import fastapi
 
-from server_utils.auth.resource_server.fastapi_dependencies import require_scopes
+from server_utils.auth.resource_server.fastapi import require_scopes
 from server_utils.auth.scopes import Scope
 from server_utils.fastapi_utils.models.json_api import (
     PydanticResponse,
