+++ robot-server/robot_server/access_control/settings/router.py
@@ -4,7 +4,7 @@ from typing import Annotated
 
 import fastapi
 
-from server_utils.auth.resource_server.fastapi_dependencies import require_scopes
+from server_utils.auth.resource_server.fastapi import require_scopes
 from server_utils.auth.scopes import Scope
 from server_utils.fastapi_utils.light_router import LightRouter
 from server_utils.fastapi_utils.models.json_api import (
