+++ robot-server/robot_server/app_setup.py
@@ -10,7 +10,7 @@ from fastapi.openapi.docs import get_redoc_html
 from fastapi.responses import HTMLResponse
 
 from opentrons import __version__
-from server_utils.auth.resource_server.fastapi_dependencies import (
+from server_utils.auth.resource_server.fastapi import (
     build_authorization_checker,
     install_authorization_checker,
 )
