+++ robot-server/robot_server/labware_offsets/router.py
@@ -7,7 +7,7 @@ from typing import Annotated, Literal
 import fastapi
 from pydantic.json_schema import SkipJsonSchema
 
-from server_utils.auth.resource_server.fastapi_dependencies import require_scopes
+from server_utils.auth.resource_server.fastapi import require_scopes
 from server_utils.auth.scopes import Scope
 from server_utils.fastapi_utils.light_router import LightRouter
 from server_utils.fastapi_utils.models.json_api import (
