+++ robot-server/robot_server/maintenance_runs/router/commands_router.py
@@ -13,7 +13,7 @@ from opentrons.protocol_engine import (
     commands as pe_commands,
 )
 from opentrons.protocol_engine.errors import CommandDoesNotExistError
-from server_utils.auth.resource_server.fastapi_dependencies import require_scopes
+from server_utils.auth.resource_server.fastapi import require_scopes
 from server_utils.auth.scopes import Scope
 from server_utils.fastapi_utils.light_router import LightRouter
 from server_utils.fastapi_utils.models.json_api import (
