+++ robot-server/robot_server/runs/router/actions_router.py
@@ -7,7 +7,7 @@ from typing import Annotated, Literal, Union
 from fastapi import Depends, status
 
 from opentrons.protocol_engine.types import DeckConfigurationType
-from server_utils.auth.resource_server.fastapi_dependencies import require_scopes
+from server_utils.auth.resource_server.fastapi import require_scopes
 from server_utils.auth.scopes import Scope
 from server_utils.fastapi_utils.light_router import LightRouter
 from server_utils.fastapi_utils.models.json_api import (
