+++ robot-server/robot_server/runs/router/commands_router.py
@@ -14,7 +14,7 @@ from opentrons.protocol_engine import (
 from opentrons.protocol_engine import (
     errors as pe_errors,
 )
-from server_utils.auth.resource_server.fastapi_dependencies import require_scopes
+from server_utils.auth.resource_server.fastapi import require_scopes
 from server_utils.auth.scopes import Scope
 from server_utils.fastapi_utils.light_router import LightRouter
 from server_utils.fastapi_utils.models.json_api import (
