+++ robot-server/robot_server/runs/router/error_recovery_policy_router.py
@@ -5,7 +5,7 @@ from typing import Annotated
 
 from fastapi import Depends, status
 
-from server_utils.auth.resource_server.fastapi_dependencies import require_scopes
+from server_utils.auth.resource_server.fastapi import require_scopes
 from server_utils.auth.scopes import Scope
 from server_utils.fastapi_utils.light_router import LightRouter
 from server_utils.fastapi_utils.models.json_api import (
