+++ robot-server/robot_server/service/tip_length/router.py
@@ -7,7 +7,7 @@ from opentrons.calibration_storage import types as cal_types
 from opentrons.calibration_storage.ot2 import models, tip_length
 from opentrons.hardware_control import API
 from opentrons_shared_data.pipette.types import LabwareUri
-from server_utils.auth.resource_server.fastapi_dependencies import require_scopes
+from server_utils.auth.resource_server.fastapi import require_scopes
 from server_utils.auth.scopes import Scope
 
 from robot_server.errors.error_responses import ErrorBody
