+++ robot-server/robot_server/system/router.py
@@ -9,7 +9,7 @@ from datetime import datetime
 
 from fastapi import Depends
 
-from server_utils.auth.resource_server.fastapi_dependencies import require_scopes
+from server_utils.auth.resource_server.fastapi import require_scopes
 from server_utils.auth.scopes import Scope
 from server_utils.fastapi_utils.light_router import LightRouter
 from server_utils.fastapi_utils.models.json_api.resource_links import (
